Step2:로또(자동) Mvc 구조로 리팩토링
