+++ src/main/kotlin/lotto/Main.kt
@@ -1,19 +1,7 @@
 package lotto
 
-import lotto.domain.Lotto
-import lotto.domain.LottoStore
-import lotto.domain.Lottos
-import lotto.view.InputView
-import lotto.view.OutputView
+import lotto.controller.LottoController
 
 fun main() {
-    val inputPrice = InputView.inputPrice()
-    val lottos = Lottos(LottoStore.buyLottos(inputPrice))
-    OutputView.printLotto(lottos)
-
-    val winningLotto = Lotto(InputView.inputWinningLotto())
-    val winningRanks = lottos.matchLotto(winningLotto)
-
-    OutputView.printWinningResult(winningRanks)
-    OutputView.printRateOfReturn(inputPrice, winningRanks)
+    LottoController.run()
 }
